back-end/requirements.txt
# FastAPI 백엔드 관련
fastapi==0.112.2
uvicorn==0.32.1
pydantic==2.10.3
starlette==0.38.2
python-dotenv==0.21.0

# 모델/데이터 분석
joblib==1.4.2
scikit-learn==1.6.1
scikit-surprise==1.1.4
pandas==2.2.3
numpy==1.26.4
scipy==1.15.1

# DB & 인증
pymongo==4.8.0
python-jose==3.3.0

idna==3.7
charset-normalizer==3.4.1
certifi==2025.1.31
urllib3==2.3.0
typing_extensions==4.12.2